refactor: 기존 Alert 로직을 AlertControllerable protocol 기본 구현으로 분리
